pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rl-3d-bin-packing"
version = "0.1.0"
description = "3D Bin Packing with Maskable PPO Reinforcement Learning"
authors = [
    {name = "Luis Garcia (Original)", email = ""},
    {name = "AI Development Team", email = ""},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8,<3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core dependencies (원본 유지)
    "gym==0.21.0",
    "nptyping==2.3.1",
    "numpy==1.21.5",
    "pillow>=9.2.0",
    "plotly>=5.9.0",
    "plotly-gif>=0.0.3",
    "pandas>=1.3.0",
    "jupyter>=1.0.0",
    
    # RL dependencies with GPU/CPU auto-detection
    "torch>=1.11.0,<2.0.0",
    "torchvision>=0.12.0",
    "stable-baselines3>=1.8.0",
    "sb3-contrib>=1.8.0",
    "tensorboard>=2.8.0",
    
    # Testing
    "pytest>=7.1.2",
]

[project.optional-dependencies]
gpu = [
    "torch>=1.11.0+cu118",
    "torchvision>=0.12.0+cu118",
]
dev = [
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pre-commit>=2.17.0",
]

[project.urls]
"Homepage" = "https://github.com/luisgarciar/3D-bin-packing"
"Repository" = "https://github.com/your-username/RL-3DbinPacking"
"Bug Tracker" = "https://github.com/your-username/RL-3DbinPacking/issues"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[project.scripts]
train-ppo = "src.train:main"
eval-model = "src.eval:main" 